Routes Refactored
